Fix: Do not use wildcards in include/Makefile.am
Wildcards are not officially supported by autotools
in Makefiles since it needs to know the exact list
of files it has to work with.

Using an absolute path was a hack that worked as long
as the path to the header files from the top source dir
was the same as the install path of those files, which
is not the case anymore.

Signed-off-by: Michael Jeanson <mjeanson@efficios.com>
Signed-off-by: Mathieu Desnoyers <mathieu.desnoyers@efficios.com>