Fix: SONAME bump to 6.1.0
In commit d6c78161aed9b2d550ce201b0a8cd5b3ee515ac8 we bumped the 'age'
part of the library version with the intention of keeping the same major
SONAME because we only introduced new symbols. However by bumping the
'age' and not the 'current' we substracted 1 to the major SONAME which
we did not intend. Seems like we missed this in testing.

Fix it by bumping the 'current' to end up with an SONAME of 6.1.0 which
is what we originally intended.

From the libtool manual for reference :

  Programs using the previous version may use the new version as drop-in
  replacement, but programs using the new version may use APIs not present
  in the previous one. In other words, a program linking against the new
  version may fail with “unresolved symbols” if linking against the old
  version at runtime: set revision to 0, bump current and age.

Signed-off-by: Michael Jeanson <mjeanson@efficios.com>
Signed-off-by: Mathieu Desnoyers <mathieu.desnoyers@efficios.com>