Remove glibc < 2.4 compat code for sched_setaffinity
Remove the rather large configure compat code for the version of
sched_setaffinity as present in glibc < 2.4.

Glibc 2.4 was released in 2006, we can safely assume nobody is still
building new systems based on an even older version.

Keep the normal sched_setaffinity detection and wrappers.

This is part of an effort to standardise our autotools setup across
project to simplify maintenance.

Change-Id: I62b1488849f88f56424f4d4ce570519d37c746c5
Signed-off-by: Michael Jeanson <mjeanson@efficios.com>
Signed-off-by: Mathieu Desnoyers <mathieu.desnoyers@efficios.com>