cleanup: remove x86 cmpxchg compat
I think the time has come to remove this compat code, i386 CPU support was
removed in the Linux kernel in 2012 [1] and the few distributions that
still build for 32bit x86 have bumped their baseline past the i386. For
example, Debian requires at least an i486 since Sarge release in 2005.

[1] https://lore.kernel.org/lkml/CAHk-=wikUaRM5H_y1Bc+QyvGi40dKDL8fnCTyz7ECbwK7aHNPQ@mail.gmail.com/

Change-Id: I8c19a6eddd43797d43a9e933099efbb3d94f4402
Signed-off-by: Michael Jeanson <mjeanson@efficios.com>
Signed-off-by: Mathieu Desnoyers <mathieu.desnoyers@efficios.com>