Fix: use underscored attributes names in external headers
Each attribute name can be used with 2 underscores before and after it.
It can be useful when another library defines a macro with the name of
an attribute.

For example, there is vpp library that defines always_inline keyword
as corresponding attribute declaration [1]. It leads to compilation
error when vpp's headers are included before urcu's ones.

According to GCC documentation [2], section 4.23,
this feature is available since, at least, GCC-2.95.3 release.

[1] https://github.com/FDio/vpp/blob/master/src/vppinfra/clib.h#L144
[2] https://gcc.gnu.org/onlinedocs/gcc-2.95.3/gcc_4.html#SEC84

Signed-off-by: Mikhail Sokolovskiy <sokolmish@gmail.com>